Create NumberQuestionTest
