+++ universal-application-tool-0.0.1/test/services/applicant/question/ApplicantQuestionTest.java
@@ -104,53 +104,6 @@ public class ApplicantQuestionTest {
   }
 
   @Test
-  public void numberQuestion_withEmptyApplicantData() {
-    ApplicantQuestion applicantQuestion =
-        new ApplicantQuestion(numberQuestionDefinition, applicantData);
-
-    assertThat(applicantQuestion.getNumberQuestion()).isInstanceOf(NumberQuestion.class);
-    assertThat(applicantQuestion.getQuestionText()).isEqualTo("question?");
-    assertThat(applicantQuestion.hasErrors()).isFalse();
-  }
-
-  @Test
-  public void numberQuestion_withPresentApplicantData() {
-    applicantData.putLong(numberQuestionDefinition.getNumberPath(), 800);
-    ApplicantQuestion applicantQuestion =
-        new ApplicantQuestion(numberQuestionDefinition, applicantData);
-    NumberQuestion numberQuestion = applicantQuestion.getNumberQuestion();
-
-    assertThat(numberQuestion.hasTypeSpecificErrors()).isFalse();
-    assertThat(numberQuestion.getNumberValue().get()).isEqualTo(800);
-  }
-
-  @Test
-  public void numberQuestion_withPresentApplicantData_failsValidation() throws Exception {
-    NumberQuestionDefinition question =
-        (NumberQuestionDefinition)
-            new QuestionDefinitionBuilder()
-                .setQuestionType(QuestionType.NUMBER)
-                .setVersion(1L)
-                .setName("question name")
-                .setPath(Path.create("applicant.my.path.name"))
-                .setDescription("description")
-                .setQuestionText(ImmutableMap.of(Locale.US, "question?"))
-                .setQuestionHelpText(ImmutableMap.of(Locale.US, "help text"))
-                .setValidationPredicates(
-                    NumberQuestionDefinition.NumberValidationPredicates.create(0, 100))
-                .build();
-    applicantData.putLong(question.getNumberPath(), 1000000);
-    ApplicantQuestion applicantQuestion = new ApplicantQuestion(question, applicantData);
-    NumberQuestion numberQuestion = applicantQuestion.getNumberQuestion();
-
-    assertThat(applicantQuestion.hasErrors()).isTrue();
-    assertThat(numberQuestion.hasTypeSpecificErrors()).isFalse();
-    assertThat(numberQuestion.getQuestionErrors())
-        .containsOnly(ValidationErrorMessage.numberTooLargeError(100));
-    assertThat(numberQuestion.getNumberValue().get()).isEqualTo(1000000);
-  }
-
-  @Test
   public void singleSelectQuestion_withEmptyApplicantData() {
     ApplicantQuestion applicantQuestion =
         new ApplicantQuestion(dropdownQuestionDefinition, applicantData);
